requirements.txt
PySide2
PyOpenssl
requests
urllib3
pillow
bs4
lxml